\name{flour}
\alias{flour}
\docType{data}
\title{Measurement of Copper Content in Wholemeal Flour}
\description{
    Measurement of copper content in wholemeal flour (measured in parts per
    million).
}
\usage{data(flour)}
\format{
    A \code{data.frame} with 24 observations (sorted in ascending order)
    on the following variables:
    \describe{
        \item{\code{copper}}{copper content \code{[double]}.}
        \item{\code{weight}}{weight \code{[double]}.}
    }
}
\details{
    The data are published in Maronna et al. (2006, p. 2).
}
\source{
Maronna, R.A., D. Martin, and V.J. Yohai (2006).
    \emph{Robust Statistics: Theory and Methods},
    John Wiley and Sons: Chichester.
}
\examples{
data(flour)
}
\keyword{datasets}
