\name{robsurvey-deprecated}
\alias{robsurvey-deprecated}

\title{Deprecated Functions in Package 'robsurvey'}

\description{
    These functions are provided for compatibility with older versions
    of the package only, and may be defunct as soon as the next release.
    \itemize{
        \item \code{\link{svyreg_huber}}
        \item \code{\link{svyreg_tukey}}
    }

    Use instead:

    \itemize{
        \item \code{\link{svyreg_huberM}}
        \item \code{\link{svyreg_tukeyM}}
    }
}
\seealso{
    \link{robsurvey-package}
}
