% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsurvey.R
\docType{package}
\name{robsurvey}
\alias{robsurvey}
\alias{robsurvey-package}
\title{robsurvey: Robust survey statistics.}
\description{
The package robsurvey is a collection of functions for robust survey
statistics.
}
\section{robsurvey functions}{

robust Horvitz-Thompson M-estimator of mean and total in \code{svymean_huber()}
and \code{svytotal_huber()}, robust trimmed Horvitz-Thompson estimator of mean
and total in \code{svymean_trimmed()} and \code{svytotal_trimmed()}, robust winsorized
Horvitz-Thompson estimator of mean and total in \code{svymean_winsorized()} and
\code{svytotal_winsorized()}, weighted median estimator in \code{weighted_median()},
weighted quantile estimator in \code{weighted_quantile()}, weighted median
absolute deviation in \code{weighted_mad()}, weighted mean and total
estimators in \code{weighted_mean()} and \code{weighted_total()}.
}

\references{
Hulliger, B. (1995). Outlier Robust Horvitz-Thompson Estimators.
Survey Methodology, 21, 79 - 87.

Hulliger, B. (2011). Main Results of the AMELI Simulation Study
on Advanced Methods for Laeken Indicators. In Proceedings of
NTTS2011, Brussels.
}
