% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsurvey.R
\name{weighted.line}
\alias{weighted.line}
\title{PROJECT refactoring the function stats::line from C to R and allow weights
AUTHORS Tobias Schoch (tobias.schoch@fhnw.ch), Beat Hulliger
Needs a weighted median function weighted.median(x,w)
Uses different quantiles for splitting the sample than line()}
\usage{
weighted.line(x, y = NULL, w, na.rm = FALSE, iter = 1)
}
\arguments{
\item{x}{the vector of the explanatory variable}

\item{y}{the vector the response varialbe}

\item{w}{the vector of the weights}

\item{iter}{number of iterations for enhancing the slop}
}
\value{
intercept and slope
}
\description{
PROJECT refactoring the function stats::line from C to R and allow weights
AUTHORS Tobias Schoch (tobias.schoch@fhnw.ch), Beat Hulliger
Needs a weighted median function weighted.median(x,w)
Uses different quantiles for splitting the sample than line()
}
\examples{
weighted.line(cars$speed, cars$dist)
weighted.line(cars$speed, cars$dist, w=rep(1:10, each=5))
}
\author{
Tobias Schoch and Beat Hulliger
}
