% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsurvey.R
\name{weighted.median.ratio}
\alias{weighted.median.ratio}
\title{Weighted robust ratio based on median}
\usage{
weighted.median.ratio(x, y = NULL, w, na.rm = FALSE)
}
\arguments{
\item{x}{is the vector of the explanatory variable, x>=0.}

\item{y}{is the vector of the response variable, y>=0.}

\item{w}{is the vector of (optional) weights, w>=0 and sum(w)>0.}

\item{na.rm}{logical TRUE when complete cases should be used}
}
\value{
a vector with two components: intercept and slope
}
\description{
A weighted median of the ratios y/x determines the slope of a
 regression through the origin.
}
\examples{
x <- c(1,2,4,5)
y <- c(1,0,5,2)
weighted.median.ratio(y~x)
}
\seealso{
[line(), weighted.line(), weighted.median.line()]
}
\author{
Beat Hulliger
}
