% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsurvey.R
\name{weighted.median}
\alias{weighted.median}
\title{Weighted lower sample median}
\usage{
weighted.median(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector whose weighted sample median is wanted}

\item{w}{a numeric vector of weights}

\item{na.rm}{a logical value indicating whether \code{NA} values should be 
stripped before the computation proceeds.}
}
\value{
weighted sample median
}
\description{
\code{weighted.median} computes the weighted lower sample median
}
\details{
See \code{\link{weighted.quantile}} for further details.
}
\examples{
x <- c(0.1, 0.35, 0.05, 0.1, 0.15, 0.05, 0.2) 
weighted.median(x, x)
}
\seealso{
\code{\link{weighted.quantile}}
}
