\name{wgt_functions}
\alias{huberWgt}
\alias{tukeyWgt}

\title{Weight functions associated with the Huber and the Tukey biweight psi-functions}
\usage{
huberWgt(x, k = 1.345)

tukeyWgt(x, k = 4.685)
}
\arguments{
\item{x}{\code{[numeric vector]} variable.}

\item{k}{\code{[scalar]} tuning constant.}
}
\value{
Numerical vector 
}
\description{
Weight functions associated with the Huber and the Tukey biweight psi-functions
} 
