% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsurvey.R
\name{wgtmeantotal}
\alias{wgtmeantotal}
\alias{weighted.total}
\alias{weighted.mean}
\title{Weighted total and mean (Horvitz-Thompson and Hajek estimators)}
\usage{
weighted.total(x, w, na.rm = FALSE)

weighted.mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{w}{a numeric vector of weights}

\item{na.rm}{a logical value indicating whether \code{NA} values
should be stripped before the computation proceeds.}
}
\value{
Estimate (scalar)
}
\description{
TBD
}
\details{
TBD
}
\note{
\code{wgtmeantotal} is a generic name for the functions documented.
}
\examples{
x <- c(0.1, 0.35, 0.05, 0.1, 0.15, 0.05, 0.2)
weighted.total(x, x)
x <- c(0.1, 0.35, 0.05, 0.1, 0.15, 0.05, 0.2)
weighted.mean(x, x)
}
