% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsurvey.R
\name{robweights}
\alias{robweights}
\title{Extraction of robustness weights (M-estimators)}
\usage{
robweights(object)
}
\arguments{
\item{object}{}
}
\value{
Vector of robustness weights
}
\description{
\code{robweights} retrieves the robustness weights from an M-estimator of class
   \code{svystat.rob}
}
\details{
Extracts the robustness weights
}
