% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsurvey.R
\name{trimwgt}
\alias{trimwgt}
\alias{weighted.mean.trimmed}
\alias{weighted.total.trimmed}
\alias{tsvymean}
\alias{tsvytotal}
\title{Weighted trimmed mean and trimmed total}
\usage{
weighted.mean.trimmed(x, w, LB = 0.05, UB = 1 - LB, na.rm = FALSE)

weighted.total.trimmed(x, w, LB = 0.05, UB = 1 - LB, na.rm = FALSE)

tsvymean(x, design, LB = 0.05, UB = 1 - LB, ...)

tsvytotal(x, design, LB = 0.05, UB = 1 - LB, ...)
}
\arguments{
\item{x}{numeric vector (\code{weighted.mean.trimmed} or \code{weighted.total.trimmed});
a formula object or variable name (\code{tsvymean} or \code{tsvytotal})}

\item{w}{numeric vector of weights}

\item{LB}{lower bound of trimming, such that \eqn{0 \leq LB < UB \leq 1}}

\item{UB}{upper bound of trimming, such that \eqn{0 \leq LB < UB \leq 1}}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}

\item{design}{a \code{survey.design} object (see \link[survey]{svydesign} in \pkg{survey})}

\item{...}{additional arguments (not used)}
}
\value{
Estimate (scalar) or object of class \code{svystat.rob}
}
\description{
Weighted trimmed estimators of the mean and total are available in two forms:
\itemize{
   \item \strong{bare-bone} functions: \code{weighted.mean.trimmed} and
 \code{weighted.total.trimmed},
   \item estimation \strong{methods}: \code{tsvymean} and
 \code{tsvytotal} (incl. variance estimation
 based on the functionality of the \pkg{survey} package).
}
}
\details{
\describe{
   \item{\emph{Overview}}{
   Robust trimmed Horvitz--Thompson total or Hajek mean
 \itemize{
 \item bare-bone functions: return the estimate (no variance estimation)
 \item estimation methods on the basis of \pkg{survey} (incl. variance estimation)
 }
   }
   \item{\emph{Variance}}{
   Variance estimates of the mean or total estimator are computed as first-order
   linearization using the design-based-estimation capabilities available
   in package \pkg{survey}.
   }
   \item{\emph{Domain estimation}}{
   Estimates for domains can be obtained using the \link[survey]{svyby}
   wrapper in the \pkg{survey} package (see examples).
   }
}
}
\note{
\code{trimwgt} is a generic name for the functions documented.
}
\section{Utility functions}{

For the methods \code{tsvymean} and \code{tsvytotal}, the following
utility functions can be used
\itemize{
   \item \code{summary} gives a summary of the estimation properties
   \item \code{\link{robweights}} retrieves the robustness weights
   \item \code{coef}, \code{vcov}, \code{residuals}, and \code{fitted}
 retrieve, respectively, the estimate, variance, residuals and fitted
 values
}
}

\examples{
library(survey)
data(api)
dstrat <- svydesign(id=~1, strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
tsvymean(~api00, dstrat, k = 2)
# Domain estimates
svyby(~api00, by = ~stype, design = dstrat, tsvymean, LB = 0.1)
}
\seealso{
\code{\link{msvymean}}, \code{\link{msvytotal}},
\code{\link{wsvymean}}, \code{\link{wsvytotal}},
\code{\link{weighted.mean.huber}}, \code{\link{weighted.total.huber}},
\code{\link{weighted.mean.winsorized}}, \code{\link{weighted.total.winsorized}}
}
