% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsurvey.R
\name{weighted.median.line}
\alias{weighted.median.line}
\title{Robust simple linear regression based on medians}
\usage{
weighted.median.line(x, y = NULL, w, type = "slopes", na.rm = FALSE)
}
\arguments{
\item{x}{is the vector of the explanatory variable}

\item{y}{is the vector of the response variable}

\item{w}{is the vector of (optional) weights}

\item{type}{is either "slopes" (default) or "products"}

\item{na.rm}{logical TRUE when complete cases should be used}
}
\value{
a vector with two components: intercept and slope
}
\description{
For type medslopes the median individual ratios response/explanatory is used
as estimator of the slope. For version ratiomeds the ratio of the median crossproduct to the
median of squares of the explanatory variable is used as the estimator of the slope.
Survey weights may be used. Missing values are neglected.
Uses utility function \code{weighted.median()}.
The median of slopes (type="slopes") uses b1=M((y-M(y,w))/(x-M(x,w)), w).
The median of crossproducts by median of squares (type="products") uses
b1=M((y-M(y,w))(x-M(x,w)), w )/ M((x-M(x,w )^2), w ), where M(x, w)
is shorthand for the function weighted.median(x, w).
The function allows weights and missing values.
}
\examples{
x <- c(1, 2, 4, 5)
y <- c(3, 2, 7, 4)
weighted.line(y~x)
weighted.median.line(y~x)
weighted.median.line(y~x, type="prod")

data(cars)
with(cars, weighted.median.line(dist ~ speed))
with(cars, weighted.median.line(dist ~ speed, type="prod"))

# weighted
w <- c(rep(1,20), rep(2,20), rep(5, 10))
with(cars, weighted.median.line(dist ~ speed, w=w))
with(cars, weighted.median.line(dist ~ speed, w=w, type="prod"))

# outlier in y
cars$dist[49] <- 360
with(cars, weighted.median.line(dist ~ speed))
with(cars, weighted.median.line(dist ~ speed, type="prod"))

# outlier in x
data(cars)
cars$speed[49] <- 72
with(cars, weighted.median.line(dist ~ speed))
with(cars, weighted.median.line(dist ~ speed, type="prod"))
}
\seealso{
[line(), weighted.line(), weighted.median.ratio()]
}
\author{
Beat Hulliger
}
