% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsurvey.R
\name{wgtmeantotal}
\alias{wgtmeantotal}
\alias{weighted.total}
\alias{weighted.mean}
\title{Weighted total and mean (Horvitz-Thompson and Hajek estimators)}
\usage{
weighted.total(x, w, na.rm = FALSE)

weighted.mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{w}{a numeric vector of weights}

\item{na.rm}{a logical value indicating whether \code{NA} values should be 
stripped before the computation proceeds.}
}
\value{
Estimate (scalar)
}
\description{
Group of functions Description section
}
\details{
Group of functions Details paragraph.
}
\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value 
sections. Also, don't need to use null, could annotate first function, and 
then using function name as the groupBy name is more intuitive.
}

