% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robsurvey.R
\name{winswgt}
\alias{winswgt}
\alias{weighted.mean.winsorized}
\alias{weighted.total.winsorized}
\alias{wsvymean}
\alias{wsvytotal}
\title{Weighted winsorized mean and trimmed total}
\usage{
weighted.mean.winsorized(x, w, LB = 0.05, UB = 1 - LB, na.rm = FALSE)

weighted.total.winsorized(x, w, LB = 0.05, UB = 1 - LB,
  na.rm = FALSE)

wsvymean(x, design, LB = 0.05, UB = 1 - LB, ...)

wsvytotal(x, design, LB = 0.05, UB = 1 - LB, ...)
}
\arguments{
\item{x}{numeric vector (\code{weighted.mean.winsorized} or
\code{weighted.total.winsorized}); a formula object or variable name
(\code{wsvymean} or \code{wsvytotal})}

\item{w}{numeric vector of weights}

\item{LB}{lower bound of winsorizing, such that \eqn{0 \leq LB < UB \leq 1}}

\item{UB}{upper bound of winsorizing, such that \eqn{0 \leq LB < UB \leq 1}}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}

\item{design}{a \code{survey.design} object (see \link[survey]{svydesign}
in \pkg{survey})}

\item{...}{additional arguments (not used)}
}
\value{
Estimate (scalar) or object of class \code{svystat.rob}
}
\description{
Weighted winsorized estimators of the mean and total are available in two forms:
\itemize{
   \item \strong{bare-bone} functions: \code{weighted.mean.winsorized} and
 \code{weighted.total.winsorized},
   \item estimation \strong{methods}: \code{wsvymean} and
 \code{wsvytotal} (incl. variance estimation
 based on the functionality of the \pkg{survey} package).
}
}
\details{
\describe{
   \item{\emph{Overview}}{
   Robust winsorized Horvitz--Thompson total or Hajek mean
 \itemize{
 \item bare-bone functions: return the estimate (no variance estimation)
 \item estimation methods on the basis of \pkg{survey} (incl. variance estimation)
 }
   }
   \item{\emph{Variance}}{
   Variance estimates of the mean or total estimator are computed as first-order
   linearization using the design-based-estimation capabilities available
   in package \pkg{survey}.
   }
   \item{\emph{Domain estimation}}{
   Estimates for domains can be obtained using the \link[survey]{svyby}
   wrapper in the \pkg{survey} package (see examples).
   }
}
}
\note{
\code{winswgt} is a generic name for the functions documented.
}
\section{Utility functions}{

For the methods \code{wsvymean} and \code{wsvytotal}, the following
utility functions can be used
\itemize{
   \item \code{summary} gives a summary of the estimation properties
   \item \code{\link{robweights}} retrieves the robustness weights
   \item \code{coef}, \code{vcov}, \code{residuals}, and \code{fitted}
 retrieve, respectively, the estimate, variance, residuals and fitted
 values
}
}

\examples{
library(survey)
data(api)
dstrat <- svydesign(id=~1, strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
wsvymean(~api00, dstrat, k = 2)
# Domain estimates
svyby(~api00, by = ~stype, design = dstrat, wsvymean, LB = 0.1)
}
\seealso{
\code{\link{msvymean}}, \code{\link{msvytotal}},
\code{\link{tsvymean}}, \code{\link{tsvytotal}},
\code{\link{weighted.mean.huber}}, \code{\link{weighted.total.huber}},
\code{\link{weighted.mean.trimmed}}, \code{\link{weighted.total.trimmed}}
}
