\name{workplace}
\alias{workplace}
\docType{data}
\title{(Modified) data from the Canadian Workplace and Employee Survey}
\description{
The \code{workplace} data.frame is a compilation Fuller (2009) 
}
\usage{data(workplace)}
\format{
   A data frame with observations on 142 workplaces, with the following variables
   \describe{
      \item{\code{ID}}{a numeric vector; identifier variable}
      \item{\code{weight}}{a numeric vector; sampling weight}
      \item{\code{employment}}{a numeric vector; (total) employment}
      \item{\code{payroll}}{a numeric vector; payroll}
   }
}
\details{
The \code{workplace} data represent a population of workplaces in the retail sector in a Canadian province. The data are not those collected by Statistics Canada, but have been generated by Fuller (2009, Example 3.1.1)  display similar characteristics to the original 1999 Canadian Workplace and Employee Survey (WES). 

\subsection{Sampling design of the 1999 Canadian Workplace and Employee Survey (WES)}{
The WES target population is defined as all workplaces operating in Canada with paid employees. The sampling frame is stratified by industry, geographic region, and size (size is defined using estimated employment). A sample of workplaces was then drawn independently in each stratum using simple random sample without replacement (sample size is determined by Neyman allocation). Several strata containing very large workplaces were sampled exhaustively; see Patak et al (1998). The original sampling weights were adjusted for nonresponse. 
}

\subsection{Remarks by Fuller (2009, p. 365)}{The original weights of WES were about 2200 for the stratum of small workplaces, about 750 for medium-sized, and about 35 for large workspaces. 
}
}

\source{
The data \code{workplace} is from Table 6.3 in Fuller (2009, pp. 366--367).
}

\references{
Fuller, W.A. (2009): Sampling Statistics, Jown Wiley \& Sons, Hoboken (NJ).
Patak, Z., Hidiroglou, M., Lavall{\'e}e, P. (1998): The methodology of the Workplace and Employee Survey, \emph{Proceedings of the Survey Research Methods Section, American Statistical Association}, pp. 83--91.  
}

\examples{
data(workplace)
}
\keyword{datasets}
